Move components into new folders for blueprint
