+++ src/components/pages/blueprints-and-devices/device-display/DeviceList.jsx
